Finished MailingListViewModel,  committing for review
