+++ Mailer/ViewModels/BaseViewModel.cs
@@ -6,7 +6,7 @@ using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Mailer
+namespace Mailer.ViewModels
 {
 	public class BaseViewModel : INotifyPropertyChanged
 	{
