+++ Mailer/ViewModels/EditAddressViewModel.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Mailer
+namespace Mailer.ViewModels
 {
 	public class EditAddressViewModel : BaseViewModel
 	{
